requirements.txt
backoff==2.2.1
comet==3.1.0
datasets==2.13.1
fairseq==0.12.2
mosestokenizer==1.2.1
msal==1.20.0
nevergrad==0.7.0
numpy==1.23.3
openai==0.27.8
PyYAML==6.0.1           # Giảm từ 6.0 để tương thích unbabel-comet
Requests==2.31.0
rouge==1.0.1
sacrebleu==2.3.1
scikit_learn==1.3.2
termcolor==2.3.0
torch==2.1.0       
tqdm==4.64.1
unbabel-comet==2.2.7    # Chỉ định rõ bản ổn định
transformers==4.44.2    # Dùng bản stable, không dùng git dev (5.0)
jsonlines==3.1.0
git+https://github.com/feralvam/easse.git
accelerate==0.21.0
bitsandbytes==0.41.1
google-generativeai==0.6.0
pyarrow==16.1.0
